be/server/requirements/requirements.txt
asgiref==3.8.1
attrs==25.3.0
cfgv==3.4.0
distlib==0.3.9
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
filelock==3.18.0
identify==2.6.9
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nodeenv==1.9.1
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
virtualenv==20.30.0
